split using bytes and strings
In python2, the "str" object is stored as bytes. It doesn't matter
which bytes or characters are stored there, we can use "str" for
both (bytes and ASCII).
But in python3 the "str" object stores Unicode code points (by
default, the prefix "b" is used to get a bytes-like object) and we
must to distinguish in which context the bytes-like object should
be used and in which string.
For clarity, file operations (read/write) in storage.py have been
swithced to binary mode, functions which work with file compression
also now work only with the bytes-like objects.

Closes #17